refactor: IssueController에서 FilterController 분리
IssueController가 이슈와 필터를 관리하는 책임을 가지고 있어 필터만을 관리하는 FilterController를 생성했습니다.
